requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
